+++ refract/__init__.py
@@ -15,5 +15,5 @@ See:
 
 from __future__ import annotations
 
-__version__ = "0.3.2.2"
+__version__ = "0.3.2.3"
 __all__ = ["__version__"]